// generated-by-analysis/Pilot8/graph-inputs/harm-quantity-with-likert-severity-data.ts
// DO NOT MODIFY (your modifications will be overwritten)
// This file is automatically (re-)generated by generate/graph-data/harm-quantity-with-likert-severity.ts.
// spell-checker: disable

	export const LostPhotos = {
	"labels": [
		"None",
		"< 1 Day",
		"Days",
		"Weeks",
		"Months",
		"1-5 Years",
		"> 5 Years"
	],
	"counts": {
		"1": [
			0,
			0,
			0,
			0,
			0,
			0,
			0
		],
		"2": [
			0,
			0,
			0,
			1,
			0,
			0,
			2
		],
		"3": [
			0,
			0,
			1,
			0,
			0,
			3,
			0
		],
		"4": [
			0,
			0,
			0,
			0,
			1,
			1,
			0
		],
		"5": [
			0,
			0,
			0,
			1,
			0,
			1,
			1
		],
		"6": [
			0,
			0,
			0,
			0,
			0,
			1,
			2
		],
		"7": [
			0,
			0,
			0,
			0,
			0,
			1,
			1
		]
	},
	"percentsOfResponses": {
		"1": [
			0,
			0,
			0,
			0,
			0,
			0,
			0
		],
		"2": [
			0,
			0,
			0,
			5,
			0,
			0,
			10
		],
		"3": [
			0,
			0,
			5,
			0,
			0,
			15,
			0
		],
		"4": [
			0,
			0,
			0,
			0,
			5,
			5,
			0
		],
		"5": [
			0,
			0,
			0,
			5,
			0,
			5,
			5
		],
		"6": [
			0,
			0,
			0,
			0,
			0,
			5,
			10
		],
		"7": [
			0,
			0,
			0,
			0,
			0,
			5,
			5
		]
	}
};
	export const LostEmails = {
	"labels": [
		"None",
		"< 1 Day",
		"Days",
		"Weeks",
		"Months",
		"1-5 Years",
		"> 5 Years"
	],
	"counts": {
		"1": [
			0,
			1,
			0,
			0,
			0,
			0,
			0
		],
		"2": [
			0,
			0,
			0,
			0,
			0,
			1,
			0
		],
		"3": [
			0,
			1,
			0,
			0,
			1,
			2,
			1
		],
		"4": [
			0,
			0,
			1,
			0,
			0,
			0,
			0
		],
		"5": [
			0,
			0,
			0,
			0,
			0,
			1,
			0
		],
		"6": [
			0,
			0,
			0,
			0,
			0,
			0,
			1
		],
		"7": [
			0,
			0,
			0,
			0,
			0,
			0,
			1
		]
	},
	"percentsOfResponses": {
		"1": [
			0,
			5,
			0,
			0,
			0,
			0,
			0
		],
		"2": [
			0,
			0,
			0,
			0,
			0,
			5,
			0
		],
		"3": [
			0,
			5,
			0,
			0,
			5,
			10,
			5
		],
		"4": [
			0,
			0,
			5,
			0,
			0,
			0,
			0
		],
		"5": [
			0,
			0,
			0,
			0,
			0,
			5,
			0
		],
		"6": [
			0,
			0,
			0,
			0,
			0,
			0,
			5
		],
		"7": [
			0,
			0,
			0,
			0,
			0,
			0,
			5
		]
	}
};