// generated-by-analysis/Pilot8/graph-inputs/account-type-data.ts
// DO NOT MODIFY (your modifications will be overwritten)
// This file is automatically (re-)generated by generate/graph-data/account-type.ts.
// spell-checker: disable

export const data = {
	"emailAccount": {
		"labels": [
			"Apple",
			"Google",
			"Microsoft",
			"Other"
		],
		"data": {
			"Compromised": {
				"Original": [
					0,
					0,
					0,
					0
				],
				"Added": [
					0,
					1,
					0,
					0
				],
				"Not Worst": [
					0,
					1,
					0,
					1
				]
			},
			"Locked Out": {
				"Original": [
					0,
					0,
					0,
					0
				],
				"Added": [
					0,
					0,
					0,
					0
				],
				"Not Worst": [
					2,
					2,
					4,
					1
				]
			}
		}
	},
	"socialAccount": {
		"labels": [
			"Facebook",
			"Instagram",
			"Twitter/X",
			"SnapChat",
			"WhatsApp",
			"WeChat",
			"Signal",
			"Other"
		],
		"data": {
			"Compromised": {
				"Original": [
					1,
					0,
					1,
					0,
					0,
					0,
					0,
					0
				],
				"Added": [
					0,
					0,
					0,
					0,
					0,
					0,
					0,
					0
				],
				"Not Worst": [
					0,
					1,
					0,
					0,
					1,
					0,
					0,
					0
				]
			},
			"Locked Out": {
				"Original": [
					0,
					0,
					0,
					0,
					0,
					0,
					0,
					0
				],
				"Added": [
					0,
					1,
					0,
					0,
					0,
					0,
					0,
					0
				],
				"Not Worst": [
					1,
					2,
					0,
					0,
					1,
					0,
					0,
					0
				]
			}
		}
	},
	"financialAccount": {
		"labels": [
			"Credit Card",
			"Credit Union",
			"Retirement",
			"Brokerage",
			"Paypal",
			"Venmo",
			"Cryptocurrency",
			"Bank",
			"Other"
		],
		"data": {
			"Compromised": {
				"Original": [
					2,
					0,
					0,
					0,
					0,
					0,
					0,
					1,
					0
				],
				"Added": [
					1,
					0,
					0,
					0,
					0,
					0,
					0,
					0,
					0
				],
				"Not Worst": [
					0,
					0,
					0,
					0,
					0,
					0,
					0,
					0,
					0
				]
			},
			"Locked Out": {
				"Original": [
					0,
					0,
					0,
					0,
					0,
					0,
					0,
					0,
					0
				],
				"Added": [
					0,
					0,
					0,
					0,
					0,
					0,
					0,
					1,
					0
				],
				"Not Worst": [
					0,
					0,
					0,
					0,
					0,
					1,
					0,
					2,
					0
				]
			}
		}
	}
};